refactor(handoff): split linked-evidence rendering into SRP submodules
`render_linked_evidence_summary` previously inlined four evidence kinds
(review packet verifier, review map, affected proof, proof plan) plus
their JSON parsers and per-kind summary structs into one ~100-line
function in output.rs. Each kind is now its own submodule under
`output/linked_evidence/`, owning its summary type, parser, and
renderer. The module root orchestrates and keeps the section header
and "linked but not readable" fallbacks.

Behavior is byte-identical: existing handoff integration tests verify
the work-order markdown is unchanged.

output.rs: 1061 -> 765 lines.
